d: add internationalisation
The D parser implements this feature similarly to the C parser,
by using gettext. gettext and dgettext are imported using extern(C).
With internationalisation enabled, YYParser.SymbolKind.toString uses GC,
because the message is parsed, saved in a string to be translated, and then
returned via the sink, as opposed to being put in the sink while it is
parsed.

* data/skeletons/d.m4 (SymbolKind.yytranslatable: New), data/skeletons/lalr1.d: Here.
* doc/bison.texi: Document it.
* tests/calc.at: Test it.