d: add token constructors support
The user can return from yylex() by calling the Symbol method of the same name as
the TokenKind reported, and adding the parameters for value and location if necessary.
These methods generate compile-time errors if the parameters are not correlated.
Token constructors work with both %union and api.value.type union.

* data/skeletons/d.m4: Here.
* tests/calc.at: Test it.